[WIP]: utilities added
